history: Move source files to internal directory
