+++ src/history/internal/HistoryEntry.cpp
