+++ src/history/internal/HistoryEntry.h
